pyproject.toml
[tool.poetry]
name = "pingpong"
version = "0.1.0"
description = ""
authors = ["Joe Nudell <jnudell@hks.harvard.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.3.5"
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
slack-sdk = "^3.21.3"
sentry-sdk = "^1.30.0"
aiocontextvars = "^0.2.2"
tiktoken = "^0.5.1"
async-throttle = "1.0.0"
azure-monitor-opentelemetry = "^1.0.0"
opentelemetry-api = "^1.20.0"
requests = "^2.31.0"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
aiohttp = "^3.9.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
pyjwt = "^2.8.0"
aiosqlite = "^0.19.0"
python-multipart = "^0.0.6"
websockets = "^12.0"
azure-communication-email = "^1.0.0"
aiosmtplib = "^3.0.1"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
discord-py = "^2.3.2"
openfga-sdk = "^0.4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pre-commit = "^3.4.0"
locust = "^2.20.0"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
